section 95 end authorization

+++ app/Http/Controllers/JobController.php
@@ -5,16 +5,20 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 
-use App\Models\Job;
+
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 use phpDocumentor\Reflection\Types\String_;
 
 
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     // @desc Show all job listings
     // @route GET /jobs
     public function index(): View
@@ -105,6 +109,9 @@ class JobController extends Controller
     // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job): string
     {
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -154,8 +161,10 @@ class JobController extends Controller
     // @route DELETE /jobs/{$id}
     public function destroy(Job $job): RedirectResponse
     {
-        //If Logo, then delete it
+        // Check if user is authorized
+        $this->authorize('delete', $job);
 
+        //If Logo, then delete it
         if ($job->company_logo) {
             Storage::delete('public/logos/' . $job->company_logo);
